requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
requests
pillow
numpy
opencv_python
ultralytics
passlib[bcrypt]
python_jose
python_multipart
httpx
realesrgan
imutils
torchreid
gdown
shapely
